# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aether-api",
    "version": "dev"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthzResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agents": {
      "post": {
        "summary": "Create Agent Route",
        "operationId": "create_agent_route_agents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "409": {
            "description": "agent_name_taken"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Agents Route",
        "operationId": "list_agents_route_agents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAgentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}": {
      "get": {
        "summary": "Get Agent Route",
        "operationId": "get_agent_route_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "agent_not_found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Agent Route",
        "operationId": "update_agent_route_agents__agent_id__put",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "agent_not_found"
          },
          "409": {
            "description": "agent_version_conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}/versions/{version}": {
      "get": {
        "summary": "Get Agent Version Route",
        "operationId": "get_agent_version_route_agents__agent_id__versions__version__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersionResponse"
                }
              }
            }
          },
          "404": {
            "description": "agent_not_found | agent_version_not_found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentDefinition": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "model": {
            "$ref": "#/components/schemas/ModelRef"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          }
        },
        "type": "object",
        "required": [
          "schema_version",
          "system_prompt"
        ],
        "title": "AgentDefinition",
        "description": "`schema_version` 을 올리는 것은 파괴적 변경 판정 대상입니다(DP-1)."
      },
      "AgentDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "definition": {
            "$ref": "#/components/schemas/AgentDefinition"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/AgentVersionEntry"
            },
            "type": "array",
            "title": "Versions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "current_version",
          "definition",
          "versions",
          "created_at",
          "updated_at"
        ],
        "title": "AgentDetailResponse",
        "description": "`GET /agents/{id}` 와 `PUT /agents/{id}` 의 성공 응답(같은 모양, spec 2.2)."
      },
      "AgentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "current_version",
          "created_at",
          "updated_at"
        ],
        "title": "AgentResponse",
        "description": "`POST /agents` 의 성공 응답."
      },
      "AgentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "current_version",
          "updated_at"
        ],
        "title": "AgentSummary",
        "description": "`GET /agents` 의 `items[]` 원소 — `created_at` 은 없습니다(spec 2.2 표 그대로)."
      },
      "AgentVersionEntry": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "version",
          "created_at"
        ],
        "title": "AgentVersionEntry",
        "description": "`GET /agents/{id}` 의 `versions[]` 원소 — 버전 번호와 생성 시각만."
      },
      "AgentVersionResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "definition": {
            "$ref": "#/components/schemas/AgentDefinition"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "version",
          "definition",
          "created_at"
        ],
        "title": "AgentVersionResponse",
        "description": "`GET /agents/{id}/versions/{version}` 의 성공 응답."
      },
      "Backoff": {
        "properties": {
          "base_seconds": {
            "type": "number",
            "title": "Base Seconds",
            "default": 0.5
          },
          "max_seconds": {
            "type": "number",
            "title": "Max Seconds",
            "default": 8.0
          }
        },
        "type": "object",
        "title": "Backoff"
      },
      "CreateAgentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "definition": {
            "$ref": "#/components/schemas/AgentDefinition"
          }
        },
        "type": "object",
        "required": [
          "name",
          "definition"
        ],
        "title": "CreateAgentRequest",
        "description": "`POST /agents` 의 요청 본문."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthzResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "service",
          "version"
        ],
        "title": "HealthzResponse",
        "description": "spec 2.3 의 계약 그대로: `{status, service, version}`."
      },
      "ListAgentsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListAgentsResponse",
        "description": "`GET /agents` 의 성공 응답."
      },
      "ModelRef": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "title": "ModelRef",
        "description": "`model.id` 가 `null` 이면 배포 설정 `AETHER_MODEL_ID` 를 씁니다."
      },
      "Policy": {
        "properties": {
          "timeout_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 120
          },
          "max_steps": {
            "type": "integer",
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 8
          },
          "model_retries": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Model Retries",
            "default": 2
          },
          "tool_retries": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Tool Retries",
            "default": 1
          },
          "backoff": {
            "$ref": "#/components/schemas/Backoff"
          }
        },
        "type": "object",
        "title": "Policy"
      },
      "UpdateAgentRequest": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/AgentDefinition"
          }
        },
        "type": "object",
        "required": [
          "definition"
        ],
        "title": "UpdateAgentRequest",
        "description": "`PUT /agents/{id}` 의 요청 본문 — `definition` 만. `name` 은 Phase 1 불변(D-9)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}